fixes imports after moving to page folders
